requirements-triposr.txt
# TripoSR 3D Generation Service Dependencies

# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# TripoSR and 3D processing
triposr==0.1.3
torch==2.1.0
torchvision==0.16.0
trimesh==4.0.5
rembg==2.0.50
pillow==10.1.0

# Utilities
numpy==1.24.3
pydantic==2.5.0
